Finishing exercise 5
Got the first few tests done no problem.  The last 2 tests however I had
a lot of issues with and no amount of googling found me the function i
was looking for.  One peek in the solutions and I found !Number.isInteger
and it was exactly what I was looking for.  I plugged it in and it worked
perfectly