Create the rest of the functions + small tweaks
Changed the add function to a much simpler a + b.  Moved the more complex
code into the sum function since it takes unlimited arguments.

Created the subtract function with a simple return a - b.

Created the power function using a build in Math.pow function

Created the factorial function by returning 1 if the
passed argument (x) is less than or equal to 1, or if x > 1 it starts a
loop which adds and multiplies all values of x together and decrements x
until it is equal to 1